# atlas-core/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/dept/delete/{id}:
        delete:
            tags:
                - SysDept
            operationId: SysDept_DeleteSysDept
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.EmptyReply'
...
    /v1/dept/deptTreeByRoleId/{roleId}:
        get:
            tags:
                - SysDept
            operationId: SysDept_GetSysRoleDept
            parameters:
                - name: roleId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.GetSysRoleDeptReply'
    /v1/dept/info/{id}:
        get:
            tags:
                - SysDept
            operationId: SysDept_GetSysDept
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.GetSysDeptReply'
    /v1/dept/list:
        post:
            tags:
                - SysDept
            operationId: SysDept_ListSysDept
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.core.v1.ListSysDeptRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.ListSysDeptReply'
    /v1/dept/list/exclude/{id}:
        get:
            tags:
                - SysDept
            operationId: SysDept_ExcludeDept
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.ListSysDeptReply'
    /v1/dept/save:
        post:
            tags:
                - SysDept
            operationId: SysDept_CreateSysDept
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.core.v1.SysDeptRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.EmptyReply'
    /v1/dept/update:
        put:
            tags:
                - SysDept
            operationId: SysDept_UpdateSysDept
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.core.v1.SysDeptRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.core.v1.EmptyReply'
...
components:
    schemas:
...
        api.core.v1.DeptReply:
            type: object
            properties:
                createdAt:
                    type: string
                remark:
                    type: string
                deptId:
                    type: string
                parentId:
                    type: string
                ancestors:
                    type: string
                deptName:
                    type: string
                orderNum:
                    type: string
                leader:
                    type: string
                phone:
                    type: string
                email:
                    type: string
                status:
                    type: string
                parentName:
                    type: string
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.core.v1.DeptReply'
            description: 部门通用返回信息
        api.core.v1.DeptTree:
            type: object
            properties:
                id:
                    type: string
                pid:
                    type: string
                label:
                    type: string
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.core.v1.DeptTree'
...
        api.core.v1.EmptyReply:
            type: object
            properties: {}
...
        api.core.v1.GetSysDeptReply:
            type: object
            properties:
                dept:
                    $ref: '#/components/schemas/api.core.v1.DeptReply'
...
        api.core.v1.GetSysRoleDeptReply:
            type: object
            properties:
                dept:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.core.v1.DeptTree'
                checkedKeys:
                    type: array
                    items:
                        type: string
...
        api.core.v1.ListSysDeptRep:
            type: object
            properties: {}
        api.core.v1.ListSysDeptReply:
            type: object
            properties:
                dept:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.core.v1.DeptReply'
...
        api.core.v1.SysDeptRep:
            type: object
            properties:
                deptId:
                    type: string
                parentId:
                    type: string
                deptName:
                    type: string
                orderNum:
                    type: string
                leader:
                    type: string
                phone:
                    type: string
                email:
                    type: string
                status:
                    type: string
...